+++ src/domain/auth-web3/domain/src/entities/web3_auth_eip4361_nonce.rs
@@ -1,0 +1,27 @@
+// Copyright Kamu Data, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+use nutype::nutype;
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+#[nutype(
+    sanitize(trim),
+    validate(len_char_min = 8),
+    derive(AsRef, Clone, Debug, PartialEq, TryFrom)
+)]
+pub struct Web3AuthenticationEip4361Nonce(String);
+
+impl Web3AuthenticationEip4361Nonce {
+    pub fn new() -> Self {
+        Self::try_new(siwe::generate_nonce()).expect("Invalid nonce generated")
+    }
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////