+++ src/domain/auth-web3/domain/src/jobs/mod.rs
@@ -1,0 +1,15 @@
+// Copyright Kamu Data, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+pub const JOB_KAMU_WEB3_AUTH_EIP4361_NONCE_SERVICE: &str =
+    "dev.kamu.web3.Web3AuthEip4361NonceService";
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////